Reorganize Add and Connection screen structure
