+++ screens/AddConnectionScreen/__tests__/SelectedChannels.js
