+++ screens/AddConnectionScreen/components/RecentConnections.js
@@ -3,31 +3,19 @@ import PropTypes from 'prop-types'
 import {
   FlatList,
   View,
-  StyleSheet,
   Text,
 } from 'react-native'
 import { findIndex } from 'lodash'
 import gql from 'graphql-tag'
 import { graphql } from 'react-apollo'
 
-import ChannelItem from '../ChannelItem'
+import ChannelItem from '../../../components/ChannelItem'
 
-const styles = StyleSheet.create({
-  container: {
-    display: 'flex',
-  },
-})
-
-class SearchConnection extends Component {
+class RecentConnections extends Component {
   constructor(props) {
     super(props)
     this.renderItem = this.renderItem.bind(this)
     this.isSelected = this.isSelected.bind(this)
-  }
-
-  shouldComponentUpdate(newProps) {
-    if (newProps.data && newProps.data.loading) { return false }
-    return true
   }
 
   isSelected(channel) {
@@ -49,13 +37,13 @@ class SearchConnection extends Component {
   }
 
   render() {
-    const { q, data } = this.props
+    const { data } = this.props
 
     if (data && data.error) {
       return (
         <View>
           <Text>
-            No results found
+            No connections found
           </Text>
         </View>
       )
@@ -64,27 +52,26 @@ class SearchConnection extends Component {
     if (data && data.loading) {
       return (
         <View>
-          <Text>Searching for {q}</Text>
+          <Text>Loading</Text>
         </View>
       )
     }
+
     return (
       <FlatList
-        contentContainerStyle={styles.container}
-        data={data.me.connection_search}
+        data={data.me.recent_connections}
         keyExtractor={item => item.id}
-        keyboardShouldPersistTaps="always"
         renderItem={this.renderItem}
       />
     )
   }
 }
 
-const ConnectionSearchQuery = gql`
-  query ConnectionSearchQuery($q: String!) {
+export const RecentConnectionsQuery = gql`
+  query RecentConnectionsQuery {
     me {
       id
-      connection_search(q: $q) {
+      recent_connections(per: 5) {
         ...ChannelThumb
       }
     }
@@ -92,19 +79,17 @@ const ConnectionSearchQuery = gql`
   ${ChannelItem.fragments.channel}
 `
 
-SearchConnection.propTypes = {
+RecentConnections.propTypes = {
   data: PropTypes.any.isRequired,
-  q: PropTypes.string.isRequired,
   onToggleConnection: PropTypes.func,
   selected: PropTypes.any,
 }
 
-SearchConnection.defaultProps = {
+RecentConnections.defaultProps = {
   onToggleConnection: () => null,
   selected: [],
 }
 
+const RecentConnectionsWithData = graphql(RecentConnectionsQuery)(RecentConnections)
 
-const ConnectionSearchWithData = graphql(ConnectionSearchQuery)(SearchConnection)
-
-export default ConnectionSearchWithData
+export default RecentConnectionsWithData
