+++ screens/AddConnectionScreen/components/SelectedChannels.js
@@ -8,8 +8,8 @@ import {
 import PropTypes from 'prop-types'
 import { Ionicons } from '@expo/vector-icons'
 
-import colors from '../../constants/Colors'
-import layout from '../../constants/Layout'
+import colors from '../../../constants/Colors'
+import layout from '../../../constants/Layout'
 
 const styles = StyleSheet.create({
   container: {
