+++ screens/AddLinkScreen/index.js
@@ -11,12 +11,12 @@ import { NavigationActions } from 'react-navigation'
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
 import { isURL } from 'validator'
 
-import FieldSet from '../../../components/FieldSet'
-import BackButton from '../../../components/BackButton'
-import HeaderRightButton from '../../../components/HeaderRightButton'
+import FieldSet from '../../components/FieldSet'
+import BackButton from '../../components/BackButton'
+import HeaderRightButton from '../../components/HeaderRightButton'
 
-import layout from '../../../constants/Layout'
-import colors from '../../../constants/Colors'
+import layout from '../../constants/Layout'
+import colors from '../../constants/Colors'
 
 const styles = StyleSheet.create({
   container: {
