+++ screens/AddTextScreen/index.js
@@ -7,12 +7,12 @@ import {
 import { NavigationActions } from 'react-navigation'
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
 
-import FieldSet from '../../../components/FieldSet'
-import BackButton from '../../../components/BackButton'
-import HeaderRightButton from '../../../components/HeaderRightButton'
+import FieldSet from '../../components/FieldSet'
+import BackButton from '../../components/BackButton'
+import HeaderRightButton from '../../components/HeaderRightButton'
 
-import layout from '../../../constants/Layout'
-import colors from '../../../constants/Colors'
+import layout from '../../constants/Layout'
+import colors from '../../constants/Colors'
 
 const styles = StyleSheet.create({
   container: {
